specify defects in electronic frame
Defects have been specified in the frame of the camera (due to the
former cameraGeom rotating all images to be in the camera frame),
which required rotating them before application, and would confuse
anyone who attempted to add new defects. Fixed this situation by
applying the rotation to the defects (using the code from isr.py
that did this) to produce new defects lists in the (unrotated)
electronics frame (as the images come off the telescope, but
including the flip in reading the raw data). Removed the rotation
in isr.py since that is no longer needed.